requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pyTelegramBotAPI==4.14.1
python-dotenv==0.20.0
requests==2.32.3
urllib3==2.4.0
